+++ docs/whats-new.rst
@@ -11,6 +11,8 @@ What's new
 - convert the note about dimension coordinates saving their units in the attributes a
   warning (:issue:`124`, :pull:`126`)
   By `Justus Magin <https://github.com/keewis>`_.
+- use ``cf-xarray``'s unit registry in the plotting example (:issue:`107`, :pull:`128`).
+  By `Justus Magin <https://github.com/keewis>`_.
 
 0.2 (May 10 2021)
 -----------------