+++ docs/whats-new.rst
@@ -14,6 +14,8 @@ What's new
 - improve the documentation on the ``format`` parameter of :py:meth:`Dataset.pint.dequantify`
   and :py:meth:`DataArray.pint.dequantify` (:issue:`121`, :pull:`127`)
   By `Justus Magin <https://github.com/keewis>`_.
+- use ``cf-xarray``'s unit registry in the plotting example (:issue:`107`, :pull:`128`).
+  By `Justus Magin <https://github.com/keewis>`_.
 
 0.2 (May 10 2021)
 -----------------