+++ docs/whats-new.rst
@@ -12,7 +12,7 @@ What's new
   warning (:issue:`124`, :pull:`126`)
   By `Justus Magin <https://github.com/keewis>`_.
 - improve the documentation on the ``format`` parameter of :py:meth:`Dataset.pint.dequantify`
-  and :py:meth:`DataArray.pint.dequantify` (:issue:`121`, :pull:`127`)
+  and :py:meth:`DataArray.pint.dequantify` (:issue:`121`, :pull:`127`, :pull:`132`)
   By `Justus Magin <https://github.com/keewis>`_.
 - use ``cf-xarray``'s unit registry in the plotting example (:issue:`107`, :pull:`128`).
   By `Justus Magin <https://github.com/keewis>`_.